requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
easyocr>=1.7.0
pymysql>=1.1.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0

